fix(stripe): wizard interaction issues (#2142)
