+++ assets/components/src/wizard/store/index.js
@@ -70,7 +70,7 @@ const actions = {
 			data: payloadPath ? get( wizardState, payloadPath ) : wizardState,
 			isQuietFetch: true,
 		} );
-		if ( ! isEmpty( updatedData ) ) {
+		if ( ! isEmpty( updatedData ) && ! updatedData.error ) {
 			return actions.setAPIDataForWizard( { slug, data: updatedData } );
 		}
 	},
@@ -99,6 +99,10 @@ const store = createReduxStore( WIZARD_STORE_NAMESPACE, {
 				isQuietLoading: Boolean( action.payload.isQuietFetch ),
 			} );
 			return apiFetch( action.payload )
+				.then( data => {
+					dispatch( WIZARD_STORE_NAMESPACE ).setError( null );
+					return data;
+				} )
 				.catch( error => {
 					dispatch( WIZARD_STORE_NAMESPACE ).setError( error );
 					return { error };
